+++ packages/react-on-rails/src/generateRenderingErrorMessage.ts
