+++ packages/react-on-rails/src/serverRenderReactComponent.ts
@@ -5,7 +5,7 @@ import type { ReactElement } from 'react';
 import createReactOutput from './createReactOutput.ts';
 import { isPromise, isServerRenderHash } from './isServerRenderResult.ts';
 import buildConsoleReplay from './buildConsoleReplay.ts';
-import handleError from './handleError.ts';
+import handleError from './generateRenderingErrorMessage.ts';
 import { renderToString } from './ReactDOMServer.cts';
 import { createResultObject, convertToError, validateComponent } from './serverRenderUtils.ts';
 import type {
